⚙️Rename Language State to Settings
